Refactor TextImputer and improve dependency management
