+++ src/shapiq/imputer/__init__.py
@@ -11,6 +11,7 @@ from .gaussian_imputer import GaussianImputer
 from .generative_conditional_imputer import GenerativeConditionalImputer
 from .marginal_imputer import MarginalImputer
 from .tabpfn_imputer import TabPFNImputer
+from .text import TextImputer
 
 __all__ = [
     "Imputer",
@@ -20,4 +21,5 @@ __all__ = [
     "TabPFNImputer",
     "GaussianImputer",
     "GaussianCopulaImputer",
+    "TextImputer"
 ]
