+++ src/shapiq/imputer/text/__init__.py
@@ -1,0 +1,11 @@
+"""Imputation strategies for handling missing feature coalitions.
+
+All imputers inherit from :class:`~shapiq.imputer.Imputer` and convert a model
+
+prediction function into a cooperative game by imputing unobserved feature values.
+
+"""
+
+from .imputer import TextImputer
+
+__all__ = ["TextImputer"]
